requirements.txt
Django==1.10.2
django-filter==0.15.2
djangorestframework==3.4.7
sqlparse==0.2.1
